implemented in_order function
Added a utility function which accepts a collections and spits it back out as a list in a specified order. This is in support of model property retrieval. A lot goes in to getting specific properties and this allows us to ensure they are in a predictable order ones we have said properties.