Added test coverage
wrote test cases for getting values (left or right) and, more importantly, getting the preferred value. The test class contains a great example of how the ModelDiff can be used to calculate preferences. Minor changes were made once testing was underway. We want to raise a KeyError, not a ValueError and 'n/a' was added as a valid preference.