Implemented model_dump replacement
Pydantic's model_dump excludes unset properties. In order to pass the compare tests that would not suffice. A new function (get_property_values) was created to match the functionality of get_property_names, but also read the model's values. As part of this work, we also gained get_values_of that will take a list of property names and fetch a dict of their values. model.compare was updated to use this new functionality in place of model_dump. Of course, test coverage was added for these new DAOModel functions.