changed set_pk argument to ignore_pk
Though another breaking change for set_values, I feel it is worth doing before finishing the beta stage. ignore_pk is easier to understand. We keep the default value as False as it makes more sense to set something to True than it is to set it to False. This also gives us a more apparent outcome. i.e. If I include a pk within the values, I would expect that to be set. With this change, that is the case unless I explicitly set it not to. Set values was also refactored to be more concise. New utility functions were added to support this: retain_indict and remove_from_dict.