Added documentation for DAO
Improved the readthedocs documentation by including a page for the dao functionality. This page covers the basic CRUD operations as well as transactions. The search features are mentioned but will be covered more on their own page. This paves the way for writing documentation for the defining models. Other cleanup and docs corrections were made. Search functionality was also added to readthedocs.