Added support for no_case_str
Added a type shortcut of no_case_str to be used on models. This will act as a str type but also indicate to the DB that the field value is case_insensitive. That means that for searches and unique checks, a lowercase value will be considered the same as an uppercase value. This is useful for fields such as 'email' where case does not matter.